agent-data-fabric/requirements.txt
# =============================================================================
# Agent Data Fabric — Python Dependencies
# =============================================================================

# Backend API
fastapi>=0.115
pydantic>=2.0
pydantic-settings>=2.0
uvicorn[standard]>=0.30

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
psycopg2-binary>=2.9
alembic>=1.13
greenlet>=3.0

# Agent Orchestration
langgraph>=0.2
langchain>=0.3
langchain-openai>=0.2
langchain-community>=0.3

# RAG
llama-index>=0.11
llama-index-vector-stores-postgres>=0.1
llama-index-embeddings-openai>=0.1
llama-index-llms-openai>=0.1
llama-index-llms-ollama>=0.1

# MCP Protocol
mcp>=1.0
fastmcp>=2.0

# Auth & Security
python-jose[cryptography]>=3.3
bcrypt>=4.0
cryptography>=42

# Azure
azure-storage-blob>=12.0
certifi>=2024.0

# Docker
docker>=7.0

# Sandbox
RestrictedPython>=7.0

# HTTP
httpx>=0.27

# Utilities
python-multipart>=0.0.9
python-dotenv>=1.0
sse-starlette>=2.0
